Fix #227

+++ src/core/cpu_mesh.rs
@@ -195,7 +195,7 @@ impl CpuMesh {
     ///
     /// Transforms the mesh by the given transformation.
     ///
-    pub fn transform(&mut self, transform: &Mat4) {
+    pub fn transform(&mut self, transform: &Mat4) -> ThreeDResult<()> {
         match self.positions {
             Positions::F32(ref mut positions) => {
                 for pos in positions.iter_mut() {
@@ -226,21 +226,27 @@ impl CpuMesh {
                 }
             }
         };
-        let normal_transform = transform.invert().unwrap().transpose();
 
-        if let Some(ref mut normals) = self.normals {
-            for n in normals.iter_mut() {
-                *n = (normal_transform * n.extend(1.0)).truncate();
+        if self.normals.is_some() || self.tangents.is_some() {
+            let normal_transform = transform
+                .invert()
+                .ok_or(CoreError::FailedInvertingTransformationMatrix)?
+                .transpose();
+
+            if let Some(ref mut normals) = self.normals {
+                for n in normals.iter_mut() {
+                    *n = (normal_transform * n.extend(1.0)).truncate();
+                }
+            }
+            if let Some(ref mut tangents) = self.tangents {
+                for t in tangents.iter_mut() {
+                    *t = (normal_transform * t.truncate().extend(1.0))
+                        .truncate()
+                        .extend(t.w);
+                }
             }
         }
-
-        if let Some(ref mut tangents) = self.tangents {
-            for t in tangents.iter_mut() {
-                *t = (normal_transform * t.truncate().extend(1.0))
-                    .truncate()
-                    .extend(t.w);
-            }
-        }
+        Ok(())
     }
 
     ///
@@ -544,17 +550,20 @@ impl CpuMesh {
     ///
     pub fn arrow(tail_length: f32, tail_radius: f32, angle_subdivisions: u32) -> Self {
         let mut arrow = Self::cylinder(angle_subdivisions);
-        arrow.transform(&Mat4::from_nonuniform_scale(
-            tail_length,
-            tail_radius,
-            tail_radius,
-        ));
+        arrow
+            .transform(&Mat4::from_nonuniform_scale(
+                tail_length,
+                tail_radius,
+                tail_radius,
+            ))
+            .unwrap();
         arrow.name = "arrow".to_string();
         let mut cone = Self::cone(angle_subdivisions);
         cone.transform(
             &(Mat4::from_translation(vec3(tail_length, 0.0, 0.0))
                 * Mat4::from_nonuniform_scale(1.0 - tail_length, 1.0, 1.0)),
-        );
+        )
+        .unwrap();
         let mut indices = arrow.indices.unwrap().into_u32();
         let cone_indices = cone.indices.unwrap().into_u32();
         let offset = indices.iter().max().unwrap() + 1;
